+++ NEWS.md
@@ -17,6 +17,8 @@
 - New argument `strict` in `$drop()` to determine whether unknown column names
   should trigger an error (#1220).
 - New method `$to_dummies()` for `DataFrame` (#1225).
+- New method `$join_where()` for `DataFrame` and `LazyFrame` to perform
+  inequality joins (#1237).
 
 ### Bug fixes